# Add caching of parsed args and keys