The PR adds caching of parsed args and keys.
That speeds up cases for expression (up to 50%) based generation like
before
```
Benchmark                            Mode  Cnt     Score     Error   Units
JmhTest._bothifyExpression           thrpt    5  1047.055 ±   5.959  ops/ms
JmhTest._letterifyExpression         thrpt    5   948.810 ± 146.551  ops/ms
JmhTest._oneMethodCallExpression     thrpt    5   666.582 ±   5.509  ops/ms
JmhTest._optionsExpression           thrpt    5   584.529 ±  12.425  ops/ms
JmhTest._regexifyExpression          thrpt    5   911.788 ±  15.694  ops/ms
JmhTest._threeMethodsCallExpression  thrpt    5   234.992 ±  11.027  ops/ms
JmhTest.csvCall                      thrpt    5    75.119 ±   2.116  ops/ms
JmhTest.jsonCall                     thrpt    5    49.711 ±   0.799  ops/ms
```
after
```
Benchmark                             Mode  Cnt     Score    Error   Units
JmhTest._bothifyExpression           thrpt    5  1914.323 ± 28.850  ops/ms
JmhTest._letterifyExpression         thrpt    5  1832.100 ± 25.612  ops/ms
JmhTest._oneMethodCallExpression     thrpt    5  1097.136 ± 18.806  ops/ms
JmhTest._optionsExpression           thrpt    5  1236.525 ± 34.222  ops/ms
JmhTest._regexifyExpression          thrpt    5  1385.327 ± 24.764  ops/ms
JmhTest._threeMethodsCallExpression  thrpt    5   294.919 ±  2.727  ops/ms
JmhTest.csvCall                      thrpt    5   100.456 ±  0.569  ops/ms
JmhTest.jsonCall                     thrpt    5    60.466 ±  1.048  ops/ms
```

test content is
```java

@BenchmarkMode(Mode.AverageTime)
@OutputTimeUnit(TimeUnit.MILLISECONDS)
@State(Scope.Benchmark)
@Fork(value = 2, jvmArgs = {"-Xms2G", "-Xmx2G"})
public class JmhTest {

    private static final Faker FAKER = new Faker();
    public static void main(String[] args) throws RunnerException {

        Options opt = new OptionsBuilder()
            .include(JmhTest.class.getSimpleName())
            .forks(1)
            .build();

        new Runner(opt).run();
    }

    @Benchmark
    @BenchmarkMode(Mode.Throughput)
    public void _optionsExpression(Blackhole blackhole) {
        blackhole.consume(FAKER.expression("#{options.option 'a','b','c','d'}"));
    }

    @Benchmark
    @BenchmarkMode(Mode.Throughput)
    public void _letterifyExpression(Blackhole blackhole) {
        blackhole.consume(FAKER.expression("#{letterify '????','true'}"));
    }

    @Benchmark
    @BenchmarkMode(Mode.Throughput)
    public void _regexifyExpression(Blackhole blackhole) {
        blackhole.consume(FAKER.expression("#{regexify '\\.\\*\\?\\+'}"));
    }

    @Benchmark
    @BenchmarkMode(Mode.Throughput)
    public void _bothifyExpression(Blackhole blackhole) {
        blackhole.consume(FAKER.expression("#{bothify '????','false'}"));
    }

    @Benchmark
    @BenchmarkMode(Mode.Throughput)
    public void _oneMethodCallExpression(Blackhole blackhole) {
        blackhole.consume(FAKER.expression("#{number.number_between '1','10'}"));
    }

    @Benchmark
    @BenchmarkMode(Mode.Throughput)
    public void _threeMethodsCallExpression(Blackhole blackhole) {
        blackhole.consume(FAKER.expression("#{Name.first_name} #{Name.first_name} #{Name.last_name}"));
    }

    @Benchmark
    @BenchmarkMode(Mode.Throughput)
    public void csvCall(Blackhole blackhole) {
        blackhole.consume(Format.toCsv(FAKER.<Name>collection()
                .suppliers(FAKER::name)
                .build())
            .headers(() -> "first_name", () -> "last_name")
            .columns(Name::firstName, Name::lastName)
            .separator(" ; ")
            .limit(10).build().get());
    }

    @Benchmark
    @BenchmarkMode(Mode.Throughput)
    public void jsonCall(Blackhole blackhole) {
        blackhole.consume(Format.toJson(
                FAKER.<Name>collection()
                    .suppliers(FAKER::name)
                    .maxLen(10)
                    .build())
            .set("firstName", Name::firstName)
            .set("lastName", Name::lastName)
            .build()
            .generate());
    }
}

```